requirements.txt
streamlit>=1.30.0
streamlit-option-menu>=0.3.6
pandas>=2.1.0
plotly>=5.18.0
firebase-admin>=6.4.0
requests>=2.31.0
python-dateutil>=2.8.2
openpyxl>=3.1.2
streamlit-authenticator>=0.2.3
PyJWT>=2.8.0
python-dotenv>=1.0.0
xlsxwriter
setuptools
pyrebase4

